Reorganise the volumes
- A single volume holds the configuration, including aliases,
  vs. a bunch of file-mounts previously.
- The spool and config are moved to separate directory trees.

This is required, as updating the aliases file atomically (i.e. by
writing to a temp file and moving it over) ends up with the mount-file
still referring to the old inode.

This means that any file edit that happened this way (e.g. by using
Ansible or a text editor outside of the container) did not appear in
the container until after its next restart.